shopify_returns_chat_agent/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.0.0
openai>=1.0.0  # optional, for LLM phase
responses>=0.25.0 